@@ -12,11 +12,15 @@
         }
     ],
     "require": {
-        "php": "^7.1",
-        "symfony/framework-bundle": "^3.4 || ^4.0",
-        "symfony/debug": "^3.4 || ^4.0",
-        "doctrine/common": "^2.3",
-        "symfony/process": "^3.4 || ^4.0"
+        "php": "^8.0",
+        "ext-pcntl": "*",
+        "symfony/framework-bundle": "^6.0",
+        "symfony/error-handler": "^6.0",
+        "doctrine/common": "^3.4",
+        "symfony/process": "^6.0",
+        "doctrine/annotations": "^2.0",
+        "symfony/cache": "^6.0",
+        "symfony/console": "^6.0"
     },
     "require-dev": {
         "doctrine/doctrine-fixtures-bundle": "*",
@@ -31,7 +35,7 @@
         "symfony/twig-bundle": "*",
         "symfony/form": "*",
         "symfony/validator": "*",
-        "phpunit/phpunit": "^6.0",
+        "phpunit/phpunit": "^9.0",
         "symfony/intl": "*"
     },
     "suggest": {
