@@ -3,30 +3,31 @@
 namespace JMS\JobQueueBundle\Tests\Functional;
 
 use Doctrine\ORM\EntityManager;
-
+use Doctrine\ORM\Tools\SchemaTool;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
+use Symfony\Component\HttpKernel\KernelInterface;
 
 class BaseTestCase extends WebTestCase
 {
-    static protected function createKernel(array $options = array())
+    protected static function createKernel(array $options = []): KernelInterface
     {
         $config = isset($options['config']) ? $options['config'] : 'default.yml';
 
         return new AppKernel($config);
     }
 
-    protected final function importDatabaseSchema()
+    final protected function importDatabaseSchema(): void
     {
         foreach (self::$kernel->getContainer()->get('doctrine')->getManagers() as $em) {
             $this->importSchemaForEm($em);
         }
     }
 
-    private function importSchemaForEm(EntityManager $em)
+    private function importSchemaForEm(EntityManager $em): void
     {
         $metadata = $em->getMetadataFactory()->getAllMetadata();
         if (!empty($metadata)) {
-            $schemaTool = new \Doctrine\ORM\Tools\SchemaTool($em);
+            $schemaTool = new SchemaTool($em);
             $schemaTool->createSchema($metadata);
         }
     }
