@@ -98,7 +98,7 @@ class CleanUpCommand extends Command
 
             $result = $con->executeQuery($incomingDepsSql, array('id' => $job->getId()));
             if ($result->fetchColumn() !== false) {
-                $em->transactional(function() use ($em, $job) {
+                $em->wrapInTransaction(function() use ($em, $job) {
                     $this->resolveDependencies($em, $job);
                     $em->remove($job);
                 });
@@ -135,7 +135,7 @@ class CleanUpCommand extends Command
             }
         }
 
-        $em->getConnection()->executeUpdate("DELETE FROM jms_job_dependencies WHERE dest_job_id = :id", array('id' => $job->getId()));
+        $em->getConnection()->executeStatement("DELETE FROM jms_job_dependencies WHERE dest_job_id = :id", array('id' => $job->getId()));
     }
 
     private function findExpiredJobs(EntityManager $em, InputInterface $input)
