@@ -18,6 +18,7 @@
 
 namespace JMS\JobQueueBundle\Tests\Entity;
 
+use JMS\JobQueueBundle\Exception\InvalidStateTransitionException;
 use JMS\JobQueueBundle\Entity\Job;
 use PHPUnit\Framework\TestCase;
 
@@ -38,10 +39,11 @@ class JobTest extends TestCase
 
     /**
      * @depends testConstruct
-     * @expectedException JMS\JobQueueBundle\Exception\InvalidStateTransitionException
      */
     public function testInvalidTransition(Job $job)
     {
+        $this->expectException(InvalidStateTransitionException::class);
+
         $job->setState('failed');
     }
 
@@ -145,12 +147,11 @@ class JobTest extends TestCase
         $this->assertSame($b, $a->getDependencies()->first());
     }
 
-    /**
-     * @expectedException \LogicException
-     * @expectedExceptionMessage You cannot add dependencies to a job which might have been started already.
-     */
     public function testAddDependencyToRunningJob()
     {
+        $this->expectException(\LogicException::class);
+        $this->expectExceptionMessage('You cannot add dependencies to a job which might have been started already.');
+
         $job = new Job('a');
         $job->setState(Job::STATE_RUNNING);
         $this->setField($job, 'id', 1);
