@@ -11,8 +11,6 @@ call_user_func(function() {
     require_once $autoloadFile;
 });
 
-\Doctrine\Common\Annotations\AnnotationRegistry::registerLoader('class_exists');
-
 use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\HttpKernel\Kernel;
@@ -37,14 +35,13 @@ class AppKernel extends Kernel
         $this->config = $config;
     }
 
-    public function registerBundles()
+    public function registerBundles(): \Traversable|array
     {
         return array(
             new \Symfony\Bundle\FrameworkBundle\FrameworkBundle(),
             new \Doctrine\Bundle\DoctrineBundle\DoctrineBundle(),
             new \Doctrine\Bundle\FixturesBundle\DoctrineFixturesBundle(),
             new \Symfony\Bundle\TwigBundle\TwigBundle(),
-
             new \JMS\JobQueueBundle\Tests\Functional\TestBundle\TestBundle(),
             new \JMS\JobQueueBundle\JMSJobQueueBundle(),
         );
@@ -55,17 +52,17 @@ class AppKernel extends Kernel
         $loader->load($this->config);
     }
 
-    public function getCacheDir()
+    public function getCacheDir(): string
     {
         return sys_get_temp_dir().'/'.Kernel::VERSION.'/JMSJobQueueBundle/'.substr(sha1($this->config), 0, 6).'/cache';
     }
 
-    public function getContainerClass()
+    public function getContainerClass(): string
     {
         return parent::getContainerClass().'_'.substr(sha1($this->config), 0, 6);
     }
 
-    public function getLogDir()
+    public function getLogDir(): string
     {
         return sys_get_temp_dir().'/'.Kernel::VERSION.'/JMSJobQueueBundle/'.substr(sha1($this->config), 0, 6).'/logs';
     }
