@@ -3,11 +3,8 @@
 namespace JMS\JobQueueBundle\Tests\Functional;
 
 use JMS\JobQueueBundle\Retry\ExponentialRetryScheduler;
-use JMS\JobQueueBundle\Retry\RetryScheduler;
 use JMS\JobQueueBundle\Tests\Functional\TestBundle\Entity\Train;
-
 use JMS\JobQueueBundle\Tests\Functional\TestBundle\Entity\Wagon;
-
 use PHPUnit\Framework\Constraint\LogicalNot;
 use Symfony\Component\EventDispatcher\EventDispatcher;
 use Doctrine\ORM\EntityManager;
@@ -253,7 +250,7 @@ class JobManagerTest extends BaseTestCase
         $this->assertEquals('terminated', $a->getState());
 
         $this->em->clear();
-        $reloadedA = $this->em->find('JMSJobQueueBundle:Job', $a->getId());
+        $reloadedA = $this->em->find(Job::class, $a->getId());
         $this->assertCount(2, $reloadedA->getRetryJobs());
     }
 
@@ -277,7 +274,7 @@ class JobManagerTest extends BaseTestCase
         $this->em->clear();
         $this->assertNotSame($defEm, $this->em);
 
-        $reloadedJ = $this->em->find('JMSJobQueueBundle:Job', $j->getId());
+        $reloadedJ = $this->em->find(Job::class, $j->getId());
 
         $reloadedWagon = $reloadedJ->findRelatedEntity('JMS\JobQueueBundle\Tests\Functional\TestBundle\Entity\Wagon');
         $reloadedWagon->state = 'broken';
@@ -287,7 +284,7 @@ class JobManagerTest extends BaseTestCase
         $this->assertTrue($defEm->contains($reloadedWagon->train));
     }
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->createClient();
         $this->importDatabaseSchema();
