@@ -2,11 +2,11 @@
 
 namespace JMS\JobQueueBundle\Tests\Functional\TestBundle\Command;
 
-use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
+use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
-class ThrowsExceptionCommand extends ContainerAwareCommand
+class ThrowsExceptionCommand extends Command
 {
     protected function configure()
     {
